requirements.txt
langchain
langchain-openai
langchain-community
langchainhub
langgraph
langchain-groq
langchain_google_community
fastapi[all]
uvicorn[standard]
pydantic
python-dotenv
sentence-transformers
faiss-cpu
pypdf
streamlit
langgraph_supervisor
langchain-huggingface
langchain_core
pytest
pytest-mock
pytest-asyncio 
pytest-cov 
pytest-env
unicorn